+++ ObjDLib/Java/src/core/chain/ZipLink.java
@@ -10,7 +10,7 @@ public class ZipLink<T, A, R> implements ChainLink<T, R> {
             @Override
             public Integer apply(final A item) {
                 if(!(ai.hasNext())) {
-                    return 1;
+                    return ((int)1);
                 } else {
                     return yield.yieldItem(ZipLink.this.f.apply(item, ai.next()));
                 }
