+++ ObjDLib/ObjC/ObjDLib/ObjDLib/Links/CNFilterLink.m
@@ -28,9 +28,9 @@ static ODClassType* _CNFilterLink_type;
 }
 
 - (CNYield*)buildYield:(CNYield*)yield {
-    return [CNYield decorateBase:yield begin:^NSInteger(NSUInteger size) {
+    return [CNYield decorateBase:yield begin:^int(NSUInteger size) {
         return [yield beginYieldWithSize:((NSUInteger)(size * _selectivity))];
-    } yield:^NSInteger(id<CNTraversable> item) {
+    } yield:^int(id<CNTraversable> item) {
         if(_predicate(item)) return [yield yieldItem:item];
         else return 0;
     }];
